requirements.txt
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry==1.6.4
flask==3.1.0
flask.cors==5.0.0
flask-wtf==1.2.2
flask-limiter==3.9.2
gunicorn==23.0.0
redis==5.2.1
